alT-Las_Project/requirements.txt
aiohttp
requests
PyQt5
opencv-python-headless
pyautogui
Pillow
mss
pybind11
pytest
coverage
